# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CaterOS SaaS API Documentation",
        "description": "Dokumentasi RESTful API untuk platform SaaS Management Catering End-to-End CaterOS.",
        "contact": {
            "name": "CaterOS Engineering Team",
            "email": "dev@cateros.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/tenant/raw-materials": {
            "get": {
                "tags": [
                    "Master Bahan Baku"
                ],
                "summary": "Daftar Master Bahan Baku",
                "description": "Get all raw materials with search and category filter.",
                "operationId": "12c14e67dd23a12a999e7a3f8b3f886e",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar bahan baku berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Master Bahan Baku"
                ],
                "summary": "Tambah Bahan Baku Baru",
                "description": "Store new raw material.",
                "operationId": "02fcc52988a34f58f04b150e821b3666",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "unit",
                                    "default_purchase_price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Daging Ayam Fillet Dada"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "RM-AYAM-01"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Daging/Unggas"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "kg"
                                    },
                                    "default_purchase_price": {
                                        "type": "number",
                                        "example": 45000
                                    },
                                    "minimum_stock": {
                                        "type": "number",
                                        "example": 5
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Supplier Pasar Induk"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bahan baku berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            }
        },
        "/tenant/raw-materials/{id}": {
            "get": {
                "tags": [
                    "Master Bahan Baku"
                ],
                "summary": "Detail Bahan Baku",
                "description": "Show raw material detail.",
                "operationId": "07bc7e08ecac4e30159e912ce4abb7ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail bahan baku berhasil diambil"
                    },
                    "404": {
                        "description": "Bahan baku tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Master Bahan Baku"
                ],
                "summary": "Update Bahan Baku & Harga Beli",
                "description": "Update raw material.",
                "operationId": "4e5cbeb2cd0454950e2a6261a7d77073",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Daging Ayam Fillet Paha"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Daging/Unggas"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "kg"
                                    },
                                    "default_purchase_price": {
                                        "type": "number",
                                        "example": 48000
                                    },
                                    "minimum_stock": {
                                        "type": "number",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bahan baku berhasil diperbarui"
                    },
                    "404": {
                        "description": "Bahan baku tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Master Bahan Baku"
                ],
                "summary": "Hapus Bahan Baku",
                "description": "Delete raw material.",
                "operationId": "e1bea87c2b6f1fbe7d39186ffbb67c0c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bahan baku berhasil dihapus"
                    },
                    "400": {
                        "description": "Bahan baku sedang digunakan dalam resep BOM"
                    },
                    "404": {
                        "description": "Bahan baku tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan Token Bearer Sanctum (contoh: 1|abcdef...)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "TenantHeader": {
                "type": "apiKey",
                "description": "Header ID Tenant yang sedang aktif (contoh: 1)",
                "name": "X-Tenant-ID",
                "in": "header"
            }
        }
    },
...
}